Lista 2

+++ lista2_exercicio.py
@@ -3,14 +3,14 @@ num2 = input("Digite outro número: ")
 sinal = input("Digite o operador(+, -, *, /): ")
 
 if sinal == "+":
-    result = int(num1) + int(num2)
+    result = float(num1) + float(num2)
     print("Resultado: ", result)
 elif sinal == "-":
-    result = int(num1) - int(num2)
+    result = float(num1) - float(num2)
     print("Resultado: ", result)
 elif sinal == "*":
-    result = int(num1) * int(num2)
+    result = float(num1) * float(num2)
     print("Resultado: ", result)
 elif sinal == "/":
-    result = int(num1) / int(num2)
+    result = float(num1) / float(num2)
     print("Resultado: ", result)
